+++ samples/Store/Backend/Favorites.fs
@@ -20,8 +20,8 @@ type Service(log, resolve, ?maxAttempts) =
     member __.Favorite(clientId, skus) =
         execute clientId (Command.Favorite(DateTimeOffset.Now, skus))
 
-    member __.Unfavorite(clientId, skus) =
-        execute clientId (Command.Unfavorite skus)
+    member __.Unfavorite(clientId, sku) =
+        execute clientId (Command.Unfavorite sku)
 
     member __.List clientId : Async<Events.Favorited []> =
         read clientId