+++ samples/Store/Integration/CartIntegration.fs
@@ -2,8 +2,6 @@
 
 open Equinox
 open Equinox.Cosmos.Integration
-open Equinox.EventStore
-open Equinox.MemoryStore
 open Swensen.Unquote
 
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)
@@ -11,16 +9,14 @@ open Swensen.Unquote
 let fold, initial = Domain.Cart.Fold.fold, Domain.Cart.Fold.initial
 let snapshot = Domain.Cart.Fold.isOrigin, Domain.Cart.Fold.snapshot
 
-let createMemoryStore () =
-    // we want to validate that the JSON UTF8 is working happily
-    VolatileStore<byte[]>()
+let createMemoryStore () = MemoryStore.VolatileStore<byte[]>()
 let createServiceMemory log store =
     Backend.Cart.Service(log, fun (id,opt) -> MemoryStore.Resolver(store, Domain.Cart.Events.codec, fold, initial).Resolve(id,?option=opt))
 
 let codec = Domain.Cart.Events.codec
 
 let resolveGesStreamWithRollingSnapshots gateway =
-    fun (id,opt) -> EventStore.Resolver(gateway, codec, fold, initial, access = AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt)
+    fun (id,opt) -> EventStore.Resolver(gateway, codec, fold, initial, access = EventStore.AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt)
 let resolveGesStreamWithoutCustomAccessStrategy gateway =
     fun (id,opt) -> EventStore.Resolver(gateway, codec, fold, initial).Resolve(id,?option=opt)